net/bnxt: Use correct hdr definition for redirect_query_tunnel_type cmd
Since tunnel_mask is 32 bit field, change logic to test for
corresponding bit set in the output of cfa_redirect_query_tunnel_type
cmd.

Bug: CTRL-31916

+++ drivers/net/bnxt/bnxt_hwrm.h
@@ -186,5 +186,5 @@ int bnxt_hwrm_tunnel_dst_port_query(struct bnxt *bp, uint8_t type,
 				uint16_t *port);
 int bnxt_hwrm_tunnel_redirect(struct bnxt *bp, uint8_t type);
 int bnxt_hwrm_tunnel_redirect_free(struct bnxt *bp, uint8_t type);
-int bnxt_hwrm_tunnel_redirect_query(struct bnxt *bp, uint8_t *type);
+int bnxt_hwrm_tunnel_redirect_query(struct bnxt *bp, uint32_t *type);
 #endif
